requirements.txt
python-telegram-bot==13.7
python-dotenv==0.21.1